book/스프링부트를 활용한 마이크로서비스 개발/kotlin/src/main/kotlin/com/example/kotlin/controller/MultiplicationResultAttemptController.kt
package com.example.kotlin.controller

import com.example.kotlin.domain.MultiplicationResultAttempt
import com.example.kotlin.service.MultiplicationService
import org.springframework.beans.factory.annotation.Autowired
import org.springframework.http.ResponseEntity
import org.springframework.web.bind.annotation.*

@RestController
@RequestMapping("/results")
class MultiplicationResultAttemptController(
    @Autowired
    val multiplicationService: MultiplicationService
) {

    @PostMapping
    fun postResult(@RequestBody multiplicationResultAttempt: MultiplicationResultAttempt): ResponseEntity<MultiplicationResultAttempt> {
        val isCorrect = multiplicationService.checkAttempt(multiplicationResultAttempt)
        val attemptCopy = MultiplicationResultAttempt(
            multiplicationResultAttempt.user, multiplicationResultAttempt.multiplication,
            multiplicationResultAttempt.resultAttempt, isCorrect
        )
        return ResponseEntity.ok(attemptCopy)
    }

    @GetMapping
    fun getStatistics(@RequestParam("alias") alias: String): ResponseEntity<List<MultiplicationResultAttempt>>{
        return ResponseEntity.ok(multiplicationService.getStatsForUser(alias))
    }
}